Implement tree block break and loot table functionality
Added features to handle block break events, specifically for tree blocks in the AquaticFarming module. The new functionality includes implementing loot tables which define drop items when a tree block is broken. Logic for tree growth and placement checks is also added to ensure trees only grow where they can be placed. This change enhances and extends the interactability and complexity of tree elements in the game.

+++ farming/plugin/src/main/kotlin/xyz/larkyy/aquaticfarming/harvestable/tree/SpawnedTree.kt
@@ -1,9 +1,13 @@
 package xyz.larkyy.aquaticfarming.harvestable.tree
 
 import com.google.gson.JsonObject
+import org.bukkit.Material
 import org.bukkit.persistence.PersistentDataType
 import xyz.larkyy.aquaticfarming.harvestable.SpawnedHarvestable
 import xyz.larkyy.aquaticfarming.harvestable.stage.TreeStage
+import xyz.larkyy.aquaticseries.AquaticSeriesLib
+import xyz.larkyy.aquaticseries.RangeAmount
+import xyz.larkyy.aquaticseries.interactable.impl.block.BlockInteractable
 import xyz.larkyy.aquaticseries.interactable.impl.block.SpawnedBlockInteractable
 import java.util.UUID
 
@@ -24,17 +28,44 @@ class SpawnedTree(
         val nextStage = nextStage() ?: return
         growthTick += 5
         if (growthTick >= nextStage.growthDuration*20) {
-            stage++
-            growthTick = 0
-            interactable.destroyBlocks()
-            interactable.despawn()
-            this.interactable = nextStage.interactable.spawn(interactable.location)
-            interactable.spawn(null, true)
-            updateData()
+            if (!canBePlaced(nextStage.interactable)) {
+                growthTick=0
+            } else {
+                stage++
+                growthTick = 0
+                interactable.destroyBlocks()
+                interactable.despawn()
+                this.interactable = nextStage.interactable.spawn(interactable.location)
+                interactable.spawn(null, true)
+                updateData()
+            }
         }
         else if (growthTick % 100 == 0) {
             updateData()
         }
+    }
+
+    fun canBePlaced(interactable: BlockInteractable): Boolean {
+        val ignoredLocations = this.interactable.associatedLocations
+
+        var canPlace = true
+        interactable.processLayerCells(interactable.shape.layers, this.interactable.location) { char, newLoc ->
+            val i = AquaticSeriesLib.INSTANCE.interactableHandler.getInteractable(newLoc.block)
+            for (ignoredLocation in ignoredLocations) {
+                if (ignoredLocation.x == newLoc.x && ignoredLocation.y == newLoc.y && ignoredLocation.z == newLoc.z) return@processLayerCells
+            }
+            if (newLoc.block.type != Material.AIR) {
+                if (interactable.shape.blocks[char] != null) {
+                    canPlace = false
+                }
+            } else if (i != null) {
+                if (!i.placeCheckEnabled) return@processLayerCells
+                if (interactable.shape.blocks[char] != null) {
+                    canPlace = false
+                }
+            }
+        }
+        return canPlace
     }
 
     fun nextStage(): TreeStage? {
